add parameter to select filepath encoding
We cannot know how what encoding was used for the filepath when the VPK file was created.
By default, we'll use utf-8, which will work for the majority of cases. For everything else,
either the app can select the encoding, or get the paths in bytes and try to guess them.

Fix #9